agentic-shared(move-resolve-provider-api-key,mechanical_provable): move the worker-side key resolution next to the launcher-side contract
resolve_provider_api_key (env var, else key file) leaves tb2_sandbox_recipe.py
unchanged and joins miles.rollout.agentic.credentials; the per-provider
wrappers in tb2_sandbox_daytona / tb2_sandbox_e2b import it from there.
